\name{evolutionR-package}
\alias{evolutionR-package}
\alias{evolutionR}

\docType{package}

\title{Evolutionary Optimization with R}

\description{Evolutionary Optimization with R}

\details{
  \tabular{ll}{
    Package: \tab evolutionR\cr
    Type: \tab Package\cr
    Version: \tab 0.1-1\cr
    Date: \tab 2014-04-17\cr
    License: \tab GPL-3\cr
  }
}

\author{
  Ronald Hochreiter \email{ron@hochreiter.net}
  
  Maintainer: Ronald Hochreiter \email{ron@hochreiter.net}
}

\references{\url{http://www.evolutionR.org/}}

\keyword{ package }

\examples{}
